Schedule v2.0 (Merged Schedule into UserShift) & add some userShift's data in sql)
