+++ src/java/controllers/UserShiftController.java
@@ -11,6 +11,7 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
 import java.util.Date;
 import java.util.List;
 import javax.servlet.ServletException;
@@ -18,6 +19,12 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import models.ShiftTime;
+import models.Users;
+import repositories.ShiftTimeRepository;
+import repositories.UsersRepository;
+import services.Utilities;
 import static services.Utilities.sdfDate;
 
 /**
@@ -47,6 +54,9 @@ public class UserShiftController extends HttpServlet {
                 case "listOf":
                     listOf(request, response);
                     break;
+                case "selectWeek":
+                    selectWeek(request, response);
+                    break;
                 case "create":
                     create(request, response);
                     break;
@@ -69,9 +79,31 @@ public class UserShiftController extends HttpServlet {
     protected void listOf(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         UserShiftRepository usr = new UserShiftRepository();
+        HttpSession session = request.getSession();
         try {
-            List<UserShift> list = usr.select();
+            List<UserShift> list = UserShiftRepository.select();
             request.setAttribute("list", list);
+            //Add for BLOCK View
+            Object selectedWeek = session.getAttribute("selectedWeek");
+            String week = Utilities.listStartEndDates().get(0);
+            if (selectedWeek != null) {
+                week = (String) selectedWeek;
+            }
+            String stringStartDate = week.substring(0, 10);
+            LocalDate startDate = Utilities.dateString(stringStartDate);
+
+            List<String> weeks = Utilities.listStartEndDates();
+            List<String> listDays = Utilities.listDaysInWeek();
+            List<LocalDate> listLocalDates = Utilities.listDatesInWeek(startDate);
+            List<LocalDate> startEndDates = Utilities.startEndDates(startDate);
+            List<ShiftTime> shifts = ShiftTimeRepository.select();
+            List<Date> listDates = Utilities.listDate(listLocalDates);
+            request.setAttribute("weeks", weeks);
+            request.setAttribute("listDays", listDays);
+            request.setAttribute("listLocalDates", listLocalDates);
+            request.setAttribute("listDates", listDates);
+            request.setAttribute("startEndDates", startEndDates);
+            request.setAttribute("shifts", shifts);
             request.getRequestDispatcher("/layouts/main.jsp").forward(request, response);
         } catch (SQLException ex) {
             ex.printStackTrace();
@@ -79,11 +111,40 @@ public class UserShiftController extends HttpServlet {
             request.getRequestDispatcher("/layouts/main.jsp").forward(request, response);
         }
     }
+    
+    protected void selectWeek(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        HttpSession session = request.getSession();
+        String op = request.getParameter("op");
+        switch (op) {
+            case "filter": {
+                try {
+//                    int weekNo = Integer.parseInt(request.getParameter("week"));
+                    String selectedWeek = request.getParameter("week");
+                    session.setAttribute("selectedWeek", selectedWeek);
+//                    request.setAttribute("week", week);
+                    response.sendRedirect(request.getContextPath() + "/userShift/listOf.do");
+                } catch (Exception ex) {
+                    request.setAttribute("message", ex.getMessage());
+                    request.setAttribute("controller", "error");
+                    request.setAttribute("action", "error");
+                    request.getRequestDispatcher("/layouts/main.jsp").forward(request, response);
+                }
+            }
+            break;
+        }
+    }
 
     protected void create(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         UserShiftRepository usr = new UserShiftRepository();
         try {
+            List<Users> userList = UsersRepository.select();
+            int shiftID = Integer.parseInt(request.getParameter("shiftID"));
+            String date = request.getParameter("date");
+            request.setAttribute("usl",userList);
+            request.setAttribute("shiftID", shiftID);
+            request.setAttribute("date", date);
             request.getRequestDispatcher("/layouts/main.jsp").forward(request, response);
         } catch (Exception ex) {
             request.setAttribute("message", ex.getMessage());
@@ -151,7 +212,7 @@ public class UserShiftController extends HttpServlet {
                     String note = request.getParameter("note");
                     boolean isOT = Boolean.parseBoolean(request.getParameter("isOT"));
                     UserShift userShift = new UserShift(userID, shiftID, date, status, note, isOT);
-                    usr.update(userShift); 
+                    usr.update(userShift);
                     response.sendRedirect(request.getContextPath() + "/userShift/listOf.do");
                 } catch (Exception ex) {
                     ex.printStackTrace();
@@ -164,9 +225,7 @@ public class UserShiftController extends HttpServlet {
             case "cancel":
                 response.sendRedirect(request.getContextPath() + "/userShift/listOf.do");
         }
-        
-            
-            
+
     }
 
     protected void delete(HttpServletRequest request, HttpServletResponse response)
